Revert change for `m_RevealCursor`, as it seems to sometimes cause errors
Not sure why (I've tried to clear my publicizer cache, etc.). It seems that, for me, Rider is refusing to acknowledge that this field is public. It still compiles fine and without any errors, however this causes annoying errors in code.

So for the sake of safety (in case anyone else is going to encounter this issue) - I've decided to just revert this specific change.

On top of that, one thing I did not consider at the time - rather than using reflection, we could have just replaced this with `te.m_RevealCursor = true` (if there was no issue after publicizing).